refactor: move rule editor component out of RuleBuiler component
