+++ app/src/components/features/rules/RuleBuilder/Body/RuleBuilderBody.css
@@ -4,7 +4,7 @@
 
 .ant-row.rule-builder-body {
   height: 80vh;
-  padding-bottom: 4rem;
+  /* padding-bottom: 4rem; */
   overflow: auto !important;
 }
