+++ app/src/views/containers/common/ContainerWithSecondarySidebar/containerWithSecondarySidebar.css
@@ -12,5 +12,5 @@
 .parent-container .outlet-container {
   flex-grow: 1;
   height: 100%;
-  overflow: auto !important;
+  overflow: hidden !important;
 }
