+++ app/src/views/features/rules/RuleEditor/components/Header/ActionButtons/CloseButton/index.js
