+++ app/src/views/features/rules/RuleEditor/components/Header/ActionButtons/CreateRuleButton/RuleInfoDialog/index.tsx
