+++ app/src/views/features/rules/RuleEditor/components/Header/ActionButtons/CreateRuleButton/RuleInfoDialog/ruleInfoDialog.css
