+++ app/src/views/features/rules/RuleEditor/components/Header/ActionButtons/CreateRuleButton/actions/insertScriptValidators.js
