+++ app/src/views/features/rules/RuleEditor/components/Header/RuleEditorHeader.css
